output: add temporary hack to prevent a mutex deadlock in the rbuf module #gklst
If PREVENT_RBUF_DEADLOCK is non-zero, allocate a small buffer to
hold the data to be streamed. This will prevent a mutex deadlock
from occuring within the rbuf module if the stream blocks on a
connection failure. This is a temporary fix - the real fix is to
set the tcp socket to non-blocking and handle it appropriately.

Signed-off-by: Michael Krufky <mkrufky@linuxtv.org>