output_stream: store pidlist subscriptions to aid in demuxing flexibility #gklst
With this changeset, one can now tune to a single transponder and send
each service and all of its associated elementary streams to separate
targets. For example, executing the following via the telnet CLI:

tune=33.1
stream=192.168.1.100:33001
tune=33.2
stream=192.168.1.101:33002

...will send each service to each respective target.

Please note: tuner / transponder / "multirec" management is not yet
robust enough to use this in a production environment.  This is just
a step in the right direction.

Signed-off-by: Michael Krufky <mkrufky@linuxtv.org>