listen: prevent resource leak if accept_socket callback not defined
prevent resource leak in socket_listen::listen_thread()
if accept_socket callback is not defined.

Also, document class socket_listen_iface:

It is the responsibility of the interface class to close
the socket accepted by accept_socket() when it is no
longer needed, otherwise it would be a resource leak.

Signed-off-by: Michael Ira Krufky <mkrufky@linuxtv.org>