Experimental: wait until all ETTs are received
We would like to display ETTs as part of the EPG output. But ETTs tend
to be transmitted asynchronously with respect to EITs. Reliably
presenting ETTs requires keeping track of which EITs have a declared
etm_location but for which the associated ETT has not arrived yet.

As a crude first implementation, we simply search for every event in
every EIT with a declared etm_location, and try to look up its ETT. It
is enabled automatically if ETT is defined and -E is supplied on the
command line.

Signed-off-by: Richard Tollerton <richard.tollerton@gmail.com>