serve: fix coverity 1435443: Unchecked return value from library
The function returns a value that indicates an error condition. If this is not checked, the error condition may not be handled correctly.

In serve_client::cmd_tuner_scan_channels_save(): Value returned from a library function is not checked for errors before being used. This value may indicate an error condition. (CWE-252)

Signed-off-by: Michael Ira Krufky <mkrufky@linuxtv.org>